# src/main/resources/api/openapi.yaml
openapi: 3.0.2
info:
  description: Logistics API provides services to find and manage entities for logistics
  title: Logistics API
  version: "1.0"
  contact:
    email: vladislavprokopenko1@gmail.com
servers:
  - url: http://localhost:8080
...
paths:
...
  /authenticationInfo:
    get:
      tags:
        - AuthenticationInfo
      summary: Find all AuthenticationInfos
      operationId: authenticationInfoFindAll
      responses:
        200:
          description: AuthenticationInfos were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuthenticationInfo'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: AuthenticationInfos were not found
        500:
          description: Internal Server Error

    post:
      tags:
        - AuthenticationInfo
      summary: Create a new AuthenticationInfo
      operationId: authenticationInfoCreate
      requestBody:
        description: AuthenticationInfo object that needs to be added to AuthenticationInfo list
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthenticationInfo'
      responses:
        201:
          description: AuthenticationInfo was created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationInfo'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        500:
          description: Internal Server Error
  /authenticationInfo/{id}:
    patch:
      tags:
        - AuthenticationInfo
      summary: Update AuthenticationInfo
      operationId: authenticationInfoUpdate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
            example: 12001
      requestBody:
        description: AuthenticationInfo object that needs to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthenticationInfo'
      responses:
        200:
          description: AuthenticationInfo was updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationInfo'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        500:
          description: Internal Server Error

    delete:
      tags:
        - AuthenticationInfo
      summary: Delete AuthenticationInfo
      operationId: authenticationInfoDelete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
            example: 12001
      responses:
        204:
          description: AuthenticationInfo was deleted successfully
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: This AuthenticationInfo id was not found
        500:
          description: Internal Server Error

    get:
      tags:
        - AuthenticationInfo
      summary: Find AuthenticationInfo by id
      operationId: authenticationInfoFindById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
            example: 12001
      responses:
        200:
          description: AuthenticationInfo was found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationInfo'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: This AuthenticationInfo was not found
        500:
          description: Internal Server Error
  /authenticationInfo/username/{username}:
    get:
      tags:
        - AuthenticationInfo
      summary: Find AuthenticationInfo by username
      operationId: authenticationInfoFindByUsername
      parameters:
        - name: username
          in: path
          description: "AuthenticationInfo username"
          required: true
          schema:
            type: string
            example: john_smith
      responses:
        200:
          description: AuthenticationInfo was found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationInfo'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: AuthenticationInfo with defined username was not found
        500:
          description: Internal Server Error
  /authenticationInfo/login:
    post:
      tags:
        - AuthenticationInfo
      summary: Send data for authorization
      operationId: authenticationInfoSend
      parameters:
        - in: header
          name: auth-token
          schema:
            type: string
      requestBody:
        description: AuthenticationInfo sending for authorization
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthenticationInfoToSend'
      responses:
        201:
          description: AuthenticationInfo was sending successfully
          content:
            text/plain:
              schema:
                description: username
                type: string
                example: john_smith
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        500:
          description: Internal Server Error
components:
  schemas:
...
    AuthenticationInfo:
      description: Object representing AuthenticationInfo
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 1201
        username:
          type: string
          example: john_smith
        password:
          type: string
          example: "{bcrypt}$2y$10$HC3hS53J4F3AYApUfV5bbewXBbV3kje5VckyZj/6QrN9RssHasDgC"
        role:
          type: string
          enum: ["ADMIN","LOGISTICIAN","DRIVER","CUSTOMER"]
          default: "LOGISTICIAN"
    AuthenticationInfoToSend:
      description: Object representing AuthenticationInfo to send
      type: object
      properties:
        username:
          type: string
          example: john_smith
        password:
          type: string
          example: "{bcrypt}$2y$10$HC3hS53J4F3AYApUfV5bbewXBbV3kje5VckyZj/6QrN9RssHasDgC"